--- docs/development/components/TableData.less
@@ -1,7 +1,0 @@
-.table-data {
-  font-size: 14px;
-
-   th, td {
-     padding: 0.6rem
-   }
-}
